fixing certain things....

+++ e-bot.py
@@ -360,7 +360,7 @@ async def mathsymbols(ctx):
 
 @client.command()
 async def lcm(ctx, x: int, y: int):
-    if type(x) is not float or type(y) is not int:
+    if type(x) is not int or type(y) is not int:
         await ctx.send("the given values are not integers bruh :|")
     else:
       await ctx.send(lcmm(x,y))
